(PC-3120) changes after review
