+++ src/components/pages/discovery-v3/utils/utils.js
@@ -44,56 +44,3 @@ export const getRecommendationSelectorByCardPosition = position => {
       return noop
   }
 }
-
-export const getOfferIdAndMediationIdAndCoordinatesApiPathQueryString = (
-  match,
-  currentRecommendation,
-  coordinates
-) => {
-  const isValid = match && typeof match === 'object' && !Array.isArray(match)
-  if (!isValid) return ''
-
-  const { params: { mediationId: pMediationId, offerId: pOfferId } = {} } = match || {}
-  const { longitude: longitude, latitude: latitude } = coordinates || {}
-
-  const offerId =
-    typeof pOfferId === 'string' &&
-    pOfferId.trim() !== '' &&
-    !DEFAULT_VIEW_IDENTIFIERS.includes(pOfferId) &&
-    pOfferId
-
-  const mediationId =
-    typeof pMediationId === 'string' &&
-    pMediationId.trim() !== '' &&
-    !DEFAULT_VIEW_IDENTIFIERS.includes(pMediationId) &&
-    pMediationId
-
-  const isSameRecoAsMatchParams =
-    currentRecommendation &&
-    currentRecommendation.mediationId === mediationId &&
-    currentRecommendation.offerId === offerId
-
-  if (isSameRecoAsMatchParams) return ''
-
-  const params = [
-    (offerId && `offerId=${offerId}`) || null,
-    (mediationId && `mediationId=${mediationId}`) || null,
-    (longitude && `longitude=${longitude}`) || null,
-    (latitude && `latitude=${latitude}`) || null,
-  ]
-
-  const query = params.filter(s => s).join('&')
-  return query
-}
-
-export const getCoordinatesApiPathQueryString = coordinates => {
-  const { longitude: longitude, latitude: latitude } = coordinates || {}
-
-  const params = [
-    (longitude && latitude && `longitude=${longitude}`) || null,
-    (longitude && latitude && `latitude=${latitude}`) || null,
-  ]
-
-  const query = params.filter(s => s).join('&')
-  return query
-}
